+++ components/Loader/Loader.tsx
@@ -1,14 +1,13 @@
 import style from './Loader.module.scss'
 
 import GridLoader from './c-grid-loader.svg'
-import Image from "next/image";
-import classNames from "classnames";
+import Image from "next/image"
 
 export const Loader = () => {
   return (
     <>
       <Image
-        className={classNames(style.loader)}
+        className={style.loader}
         src={GridLoader} aria-hidden='true'
       />
     </>
