+++ pages/guide/chapters/ch-four/ch-four.module.scss
@@ -1,0 +1,8 @@
+@import "../../../../styles/variables";
+
+.isActive {
+  color: $green;
+}
+.isNotActive {
+  color: $red;
+}
