--- pages/guide/chapters/index.ts
@@ -1,10 +1,0 @@
-export * from './ch-one/index'
-export * from './ch-two/index'
-export * from './ch-three/index'
-export * from './ch-four/index'
-export * from './ch-five/ch-five'
-export * from './ch-six/index'
-export * from './ch-seven/index'
-export * from './ch-eight/index'
-export * from './ch-nine/index'
-export * from './ch-ten/index'
