+++ styles/main.scss
@@ -1,4 +1,5 @@
 @import "variables";
+@import "functions";
 
 html,
 body {
@@ -52,6 +53,7 @@ ul {
 
   li {
     padding: .7rem 1rem;
+    font-size: 1.1rem;
 
     a {
       display: block;
@@ -86,7 +88,6 @@ img {
 }
 
 code {
-  display: block;
   font-size: 1.4rem;
 }
 
@@ -145,3 +146,4 @@ select {
 .has-padding {
   padding: 1rem;
 }
+
